chore: moved tests using out-of-line module
